+++ lib/rails-api/application.rb
@@ -1,5 +1,6 @@
 require 'rails/version'
 require 'rails/application'
+require 'rails-api/public_exceptions'
 
 module Rails
   class Application < Engine
@@ -26,7 +27,7 @@ module Rails
         middleware.use ::Rack::Runtime
         middleware.use ::ActionDispatch::RequestId
         middleware.use ::Rails::Rack::Logger, config.log_tags # must come after Rack::MethodOverride to properly log overridden methods
-        middleware.use ::ActionDispatch::ShowExceptions, config.exceptions_app || ActionDispatch::PublicExceptions.new(Rails.public_path)
+        middleware.use ::ActionDispatch::ShowExceptions, config.exceptions_app || Rails::API::PublicExceptions.new
         middleware.use ::ActionDispatch::DebugExceptions
         middleware.use ::ActionDispatch::RemoteIp, config.action_dispatch.ip_spoofing_check, config.action_dispatch.trusted_proxies