+++ test/api_controller/renderers_test.rb
@@ -12,12 +12,18 @@ class Model
 end
 
 class RenderersApiController < ActionController::API
+  use ActionDispatch::ShowExceptions, Rails::API::PublicExceptions.new
+
   def one
     render :json => Model.new
   end
 
   def two
     render :xml => Model.new
+  end
+
+  def boom
+    raise "boom"
   end
 end
 
@@ -34,5 +40,31 @@ class RenderersApiTest < ActionController::TestCase
     get :two
     assert_response :success
     assert_equal({ :a => 'b' }.to_xml, @response.body)
+  end
+end
+
+class RenderExceptionsTest < ActionDispatch::IntegrationTest
+  def setup
+    @app = RenderersApiController.action(:boom)
+  end
+
+  def test_render_json_exception
+    get "/fake", {}, 'HTTP_ACCEPT' => 'application/json'
+    assert_response :internal_server_error
+    assert_equal 'application/json', response.content_type.to_s
+    assert_equal({ :status => '500', :error => 'boom' }.to_json, response.body)
+  end
+
+  def test_render_xml_exception
+    get "/fake", {}, 'HTTP_ACCEPT' => 'application/xml'
+    assert_response :internal_server_error
+    assert_equal 'application/xml', response.content_type.to_s
+    assert_equal({ :status => '500', :error => 'boom' }.to_xml, response.body)
+  end
+
+  def test_render_fallback_exception
+    get "/fake", {}, 'HTTP_ACCEPT' => 'text/csv'
+    assert_response :internal_server_error
+    assert_equal 'text/html', response.content_type.to_s
   end
 end